add work and break timer options
